# FE/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Sports Center Management API",
    "version": "1.0.0",
    "description": "Backend API for Sports Center Management System – Flow 1 (User & Membership), Flow 2 (Class & Schedule), Flow 3 (Payment & Report)"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Development Server"
    },
    {
      "url": "https://sports-center-management-system.onrender.com/api/v1",
      "description": "Production Server (Render)"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Missing, invalid or expired token",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "example": {
                "success": false,
                "message": "Unauthorized: invalid or expired token"
              }
            }
          }
        }
      },
      "Forbidden": {
        "description": "Insufficient role permissions",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "example": {
                "success": false,
                "message": "Forbidden: insufficient permissions"
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "example": {
                "success": false,
                "message": "Record not found"
              }
            }
          }
        }
      },
...
      "ServerError": {
        "description": "Unexpected internal server error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "example": {
                "success": false,
                "message": "Internal server error"
              }
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/notifications": {
      "get": {
        "summary": "Get my notifications",
        "tags": [
          "Notifications"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string",
              "enum": [
                "MEMBER_REGISTERED",
                "CHAT_MESSAGE",
                "SUBSCRIPTION_EXPIRING",
                "SUBSCRIPTION_EXPIRED",
                "SUBSCRIPTION_CANCELLED",
                "UPCOMING_CLASS",
                "SCHEDULE_CANCELLED",
                "SCHEDULE_UPDATED",
                "ENROLLMENT_CONFIRMED",
                "ENROLLMENT_CANCELLED",
                "TRAINING_PLAN_ASSIGNED",
                "NEW_CLASS",
                "COACH_CHANGED",
                "PAYMENT_SUCCESS",
                "PAYMENT_REFUNDED",
                "PAYMENT_FAILED"
              ]
            },
            "description": "Filter by notification type"
          },
          {
            "in": "query",
            "name": "isRead",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            },
            "description": "Filter by read status"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of notifications"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/notifications/unread-count": {
      "get": {
        "summary": "Get count of unread notifications",
        "tags": [
          "Notifications"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Unread count retrieved successfully"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/notifications/mark-all-read": {
      "patch": {
        "summary": "Mark all my notifications as read",
        "tags": [
          "Notifications"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "All notifications marked as read"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/notifications/{id}/read": {
      "patch": {
        "summary": "Mark a specific notification as read",
        "tags": [
          "Notifications"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Notification ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Notification marked as read"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/notifications/trigger-upcoming-reminders": {
      "post": {
        "summary": "Manually trigger upcoming class reminders (within 24h)",
        "tags": [
          "Notifications"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Reminders triggered successfully"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
  }
}